refactor: Renamed ListCheck and TextCheck to ...Evaluate
