+++ src/SmartHopper.Components/Properties/Resources.resx
@@ -127,8 +127,8 @@
   <data name="listalter" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\listalter.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
-  <data name="listcheck" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Resources\listcheck.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  <data name="listevaluate" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\listevaluate.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
   <data name="listfilter" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\listfilter.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
@@ -136,8 +136,8 @@
   <data name="textgenerate" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\textgenerate.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
-  <data name="textcheck" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Resources\textcheck.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  <data name="textevaluate" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\Resources\textevaluate.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
   <data name="textalter" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\Resources\textalter.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
